IterativeSolvers.jl extension

GridapSolvers provides wrappers for some iterative solvers from the package IterativeSolvers.jl:

IterativeSolversExt.IterativeLinearSolver — Type

struct IterativeLinearSolver <: LinearSolver
  ...
end

Wrappers for IterativeSolvers.jl krylov-like iterative solvers.

All wrappers take the same kwargs as the corresponding solver in IterativeSolvers.jl.

The following solvers are available:

IS_ConjugateGradientSolver

IS_GMRESSolver

IS_MINRESSolver

IS_SSORSolver

GridapSolvers.IS_ConjugateGradientSolver — Function

IS_ConjugateGradientSolver(;kwargs...)

Wrapper for the Conjugate Gradient solver.

GridapSolvers.IS_GMRESSolver — Function

IS_GMRESSolver(;kwargs...)

Wrapper for the GMRES solver.

GridapSolvers.IS_MINRESSolver — Function

IS_MINRESSolver(;kwargs...)

Wrapper for the MINRES solver.

GridapSolvers.IS_SSORSolver — Function

IS_SSORSolver(ω;kwargs...)

Wrapper for the SSOR solver.