GridapSolvers.MultilevelTools

Nested subpartitions

One of the main difficulties of multilevel algorithms is dealing with the complexity of having multiple subcommunicators. We provide some tools to deal with it. In particular we introduce HierarchicalArrays.

GridapSolvers.MultilevelTools.generate_level_parts — Function

generate_level_parts(root_parts::AbstractArray,num_procs_x_level::Vector{<:Integer})

From a root communicator root_parts, generate a sequence of nested subcommunicators with sizes given by num_procs_x_level.

GridapSolvers.MultilevelTools.HierarchicalArray — Type

HierarchicalArray{T,A,B} <: AbstractVector{T}

Array of hierarchical (nested) distributed objects. Each level might live in a different subcommunicator. If a processor does not belong to subcommunicator ranks[i], then array[i] is nothing.

However, it assumes: - The subcommunicators are nested, so that ranks[i] contains ranks[i+1]. - The first subcommunicator does not have empty parts.

Base.map — Function

Base.map(f::Function,args::Vararg{HierarchicalArray,N}) where N

Maps a function to a set of HierarchicalArrays. The function is applied only in the subcommunicators where the processor belongs to.

GridapSolvers.MultilevelTools.with_level — Function

with_level(f::Function,a::HierarchicalArray,lev::Integer;default=nothing)

Applies a function to the lev-th level of a HierarchicalArray. If the processor does not belong to the subcommunicator of the lev-th level, then default is returned.

ModelHierarchies and FESpaceHierarchies

This objects are the multilevel counterparts of Gridap's DiscreteModel and FESpace.

GridapSolvers.MultilevelTools.ModelHierarchy — Type

const ModelHierarchy = HierarchicalArray{<:ModelHierarchyLevel}

A ModelHierarchy is a hierarchical array of ModelHierarchyLevel objects. It stores the adapted/redistributed models and the corresponding subcommunicators.

For convenience, implements some of the API of DiscreteModel.

GridapSolvers.MultilevelTools.ModelHierarchyLevel — Type

Single level for a ModelHierarchy.

Note that model_red and red_glue might be of type Nothing whenever there is no redistribution in a given level.

ref_glue is of type Nothing on the coarsest level.

GridapSolvers.MultilevelTools.CartesianModelHierarchy — Function

CartesianModelHierarchy(
  ranks,np_per_level,domain,nc::NTuple{D,<:Integer};
  num_refs_coarse::Integer = 0,
  add_labels!::Function = (labels -> nothing),
  map::Function = identity,
  isperiodic::NTuple{D,Bool} = Tuple(fill(false,D))
) where D

Returns a ModelHierarchy with a Cartesian model as coarsest level. The i-th level will be distributed among np_per_level[i] processors. The seed model is given by cmodel = CartesianDiscreteModel(domain,nc).

GridapSolvers.MultilevelTools.FESpaceHierarchy — Type

const FESpaceHierarchy = HierarchicalArray{<:FESpaceHierarchyLevel}

A FESpaceHierarchy is a hierarchical array of FESpaceHierarchyLevel objects. It stores the adapted/redistributed fe spaces and the corresponding subcommunicators.

For convenience, implements some of the API of FESpace.

Grid transfer operators

To move information between different levels, we will require grid transfer operators. Although any custom-made operator can be used, we provide some options.

GridapSolvers.MultilevelTools.DistributedGridTransferOperator — Type

GridapSolvers.MultilevelTools.RestrictionOperator — Function

GridapSolvers.MultilevelTools.ProlongationOperator — Function

GridapSolvers.MultilevelTools.MultiFieldTransferOperator — Type

Misc

GridapSolvers.MultilevelTools.LocalProjectionMap — Type

struct LocalProjectionMap <: Map

Map that projects a field/field-basis onto another local reference space given by a ReferenceFE.

Example:

```julia model = CartesianDiscreteModel((0,1,0,1),(2,2))

reffeh1 = ReferenceFE(QUAD,lagrangian,Float64,1,space=:Q) reffel2 = ReferenceFE(QUAD,lagrangian,Float64,1,space=:P) U = FESpace(model,reffeh1) uh1 = interpolate(f,U)

qdegree = 2 Π = LocalProjectionMap(reffel2,qdegree) ul2 = Π(u_h1) ```