GridapSolvers.PatchBasedSmoothers

GridapSolvers.PatchBasedSmoothers.BlockJacobiProlongationOperator — Type

GridapSolvers.PatchBasedSmoothers.BlockJacobiProlongationOperator — Method

GridapSolvers.PatchBasedSmoothers.PatchProlongationOperator — Type

struct PatchProlongationOperator end

A PatchProlongationOperator is a modified prolongation operator such that given a coarse solution xH, it maps it to a fine mesh solution xh given by

xh = Ih(xH) - yh

where yh is a subspace-based correction computed by solving local problems on coarse cell patches within the fine mesh.

GridapSolvers.PatchBasedSmoothers.PatchProlongationOperator — Method

function PatchProlongationOperator(
  lev   :: Integer,
  sh    :: FESpaceHierarchy,
  ptopo :: PatchTopology,
  lhs   :: Function,
  rhs   :: Function;
  is_nonlinear=false,
  collect_factorizations=false
)

Returns an instance of PatchProlongationOperator for a given level lev and a given FESpaceHierarchy sh. The subspace-based correction on a solution uH is computed by solving local problems given by

  lhs(u_i,v_i) = rhs(uH,v_i)  ∀ v_i ∈ V_i

where V_i is a restriction of sh[lev] onto the patches given by the PatchTopology ptopo.

GridapSolvers.PatchBasedSmoothers.CoarsePatchTopology — Function

CoarsePatchTopology(model::AdaptedDiscreteModel)

Given an AdaptedDiscreteModel, returns a PatchTopology for the fine model such that each patch corresponds to a coarse cell in the parent model.