GridapSolvers.LinearSolvers

Krylov solvers

GridapSolvers.LinearSolvers.CGSolver — Type

struct CGSolver <: LinearSolver
  ...
end

CGSolver(Pl;maxiter=1000,atol=1e-12,rtol=1.e-6,flexible=false,verbose=0,name="CG")

Left-Preconditioned Conjugate Gradient solver.

GridapSolvers.LinearSolvers.MINRESSolver — Type

struct MINRESSolver <: LinearSolver 
  ...
end

MINRESSolver(m;Pr=nothing,Pl=nothing,maxiter=100,atol=1e-12,rtol=1.e-6,verbose=false,name="MINRES")

MINRES solver, with optional right and left preconditioners Pr and Pl.

GridapSolvers.LinearSolvers.GMRESSolver — Type

struct GMRESSolver <: LinearSolver 
  ...
end

GMRESSolver(m;Pr=nothing,Pl=nothing,restart=false,m_add=1,maxiter=100,atol=1e-12,rtol=1.e-6,verbose=false,name="GMRES")

GMRES solver, with optional right and left preconditioners Pr and Pl.

The solver starts by allocating a basis of size m. Then:

If restart=true, the basis size is fixed and restarted every m iterations.

If restart=false, the basis size is allowed to increase. When full, the solver allocates m_add new basis vectors.

GridapSolvers.LinearSolvers.FGMRESSolver — Type

struct FGMRESSolver <: LinearSolver 
  ...
end

FGMRESSolver(m,Pr;Pl=nothing,restart=false,m_add=1,maxiter=100,atol=1e-12,rtol=1.e-6,verbose=false,name="FGMRES")

Flexible GMRES solver, with right-preconditioner Pr and optional left-preconditioner Pl.

The solver starts by allocating a basis of size m. Then:

If restart=true, the basis size is fixed and restarted every m iterations.

If restart=false, the basis size is allowed to increase. When full, the solver allocates m_add new basis vectors.

GridapSolvers.LinearSolvers.krylov_mul! — Function

Computes the Krylov matrix-vector product

y = Pl⁻¹⋅A⋅Pr⁻¹⋅x

by solving

Pr⋅wr = x wl = A⋅wr Pl⋅y = wl

GridapSolvers.LinearSolvers.krylov_residual! — Function

Computes the Krylov residual

r = Pl⁻¹(A⋅x - b)

by solving

w = A⋅x - b Pl⋅r = w

Smoothers

Given a linear system $Ax = b$, a smoother is an operator S that takes an iterative solution $x_k$ and its residual $r_k = b - A x_k$, and modifies them in place

\[ S : (x_k,r_k) \rightarrow (x_{k+1},r_{k+1})\]

such that $|r_{k+1}| < |r_k|$.

GridapSolvers.LinearSolvers.RichardsonSmoother — Type

struct RichardsonSmoother{A} <: LinearSolver
  M     :: A
  niter :: Int64
  ω     :: Float64
end

Iterative Richardson smoother. Given a solution x and a residual r, performs niter Richardson iterations with damping parameter ω using the linear solver M. A Richardson iteration is given by:

dx = ω * inv(M) * r
x  = x + dx
r  = r - A * dx

Updates both the solution x and the residual r in place.

GridapSolvers.LinearSolvers.RichardsonSmoother — Method

function RichardsonSmoother(M::LinearSolver,niter::Int=1,ω::Float64=1.0)

Returns an instance of RichardsonSmoother from its underlying properties.

Preconditioners

Given a linear system $Ax = b$, a preconditioner is an operator that takes an iterative residual $r_k$ and returns a correction $dx_k$.

GridapSolvers.LinearSolvers.JacobiLinearSolver — Type

struct JacobiLinearSolver <: LinearSolver end

Given a matrix A, the Jacobi or Diagonal preconditioner is defined as P = diag(A).

GridapSolvers.LinearSolvers.GMGLinearSolverFromMatrices — Type

struct GMGLinearSolverFromMatrices <: LinearSolver
  ...
end

Geometric MultiGrid solver, from algebraic parts.

GridapSolvers.LinearSolvers.GMGLinearSolverFromWeakform — Type

struct GMGLinearSolverFromWeakForm <: LinearSolver
  ...
end

Geometric MultiGrid solver, from FE parts.

GridapSolvers.LinearSolvers.GMGLinearSolver — Function

GMGLinearSolver(
  mh::ModelHierarchy,
  matrices::AbstractArray{<:AbstractMatrix},
  prolongations,
  restrictions;
  pre_smoothers   = Fill(RichardsonSmoother(JacobiLinearSolver(),10),num_levels(mh)-1),
  post_smoothers  = pre_smoothers,
  coarsest_solver = LUSolver(),
  mode::Symbol    = :preconditioner,
  maxiter = 100, atol = 1.0e-14, rtol = 1.0e-08, verbose = false,
)

Creates an instance of GMGLinearSolverFromMatrices from the underlying model hierarchy, the system matrices at each level and the transfer operators and smoothers at each level except the coarsest.

The solver has two modes of operation, defined by the kwarg mode:

:solver: The GMG solver takes a rhs b and returns a solution x.

:preconditioner: The GMG solver takes a residual r and returns a correction dx.

GMGLinearSolver(
  mh::ModelHierarchy,
  trials::FESpaceHierarchy,
  tests::FESpaceHierarchy,
  biforms::AbstractArray{<:Function},
  interp,
  restrict;
  pre_smoothers   = Fill(RichardsonSmoother(JacobiLinearSolver(),10),num_levels(mh)-1),
  post_smoothers  = pre_smoothers,
  coarsest_solver = Gridap.Algebra.LUSolver(),
  mode::Symbol    = :preconditioner,
  is_nonlinear    = false,
  maxiter = 100, atol = 1.0e-14, rtol = 1.0e-08, verbose = false,
)

Creates an instance of GMGLinearSolverFromMatrices from the underlying model hierarchy, the trial and test FEspace hierarchies, the weakform lhs at each level and the transfer operators and smoothers at each level except the coarsest.

The solver has two modes of operation, defined by the kwarg mode:

:solver: The GMG solver takes a rhs b and returns a solution x.

:preconditioner: The GMG solver takes a residual r and returns a correction dx.

Wrappers

PETSc

Building on top of GridapPETSc.jl, GridapSolvers provides specific solvers for some particularly complex PDEs:

GridapSolvers.LinearSolvers.ElasticitySolver — Type

struct ElasticitySolver <: LinearSolver
  ...
end

GMRES + AMG solver, specifically designed for linear elasticity problems.

Follows PETSc's documentation for PCAMG and MatNullSpaceCreateRigidBody.

GridapSolvers.LinearSolvers.ElasticitySolver — Method

function ElasticitySolver(space::FESpace; maxiter=500, atol=1.e-12, rtol=1.e-8)

Returns an instance of ElasticitySolver from its underlying properties.

GridapSolvers.LinearSolvers.CachedPETScNS — Type

struct CachedPETScNS <: NumericalSetup
  ...
end

Wrapper around a PETSc NumericalSetup, providing highly efficiend reusable caches:

When converting julia vectors/PVectors to PETSc vectors, we purposely create aliasing of the vector values. This means we can avoid copying data from one to another before solving, but we need to be careful about it.

This structure takes care of this, and makes sure you do not attempt to solve the system with julia vectors that are not the ones you used to create the solver cache.

GridapSolvers.LinearSolvers.CachedPETScNS — Method

function CachedPETScNS(ns::PETScLinearSolverNS,x::AbstractVector,b::AbstractVector)

Create a new instance of CachedPETScNS from its underlying properties. Once this structure is created, you can only solve the system with the same vectors you used to create it.

GridapSolvers.LinearSolvers.get_dof_coordinates — Function

get_dof_coordinates(space::FESpace)

Given a lagrangian FESpace, returns the physical coordinates of the DoFs, as required by some PETSc solvers. See PETSc documentation.

IterativeSolvers.jl

GridapSolvers provides wrappers for some iterative solvers from the package IterativeSolvers.jl:

GridapSolvers.LinearSolvers.IterativeLinearSolver — Type

struct IterativeLinearSolver <: LinearSolver
  ...
end

Wrappers for IterativeSolvers.jl krylov-like iterative solvers.

All wrappers take the same kwargs as the corresponding solver in IterativeSolvers.jl.

The following solvers are available:

IS_ConjugateGradientSolver

IS_GMRESSolver

IS_MINRESSolver

IS_SSORSolver

GridapSolvers.LinearSolvers.IS_ConjugateGradientSolver — Function

IS_ConjugateGradientSolver(;kwargs...)

Wrapper for the Conjugate Gradient solver.

GridapSolvers.LinearSolvers.IS_GMRESSolver — Function

IS_GMRESSolver(;kwargs...)

Wrapper for the GMRES solver.

GridapSolvers.LinearSolvers.IS_MINRESSolver — Function

IS_MINRESSolver(;kwargs...)

Wrapper for the MINRES solver.

GridapSolvers.LinearSolvers.IS_SSORSolver — Function

IS_SSORSolver(ω;kwargs...)

Wrapper for the SSOR solver.